pyproject.toml
[tool.poetry]
name = "totr"
version = "0.1.0"
description = ""
authors = ["Noppanat Wadlom <noppanat.wad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
elasticsearch = {version = "7.9.1", extras = ["async"]}
python-dotenv = "^1.0.1"
phx-class-registry = "^5.1.1"
backoff = "^2.2.1"
openai = "^1.65.2"
transformers = {extras = ["torch"], version = "^4.49.0"}
numpy = "<2.0"
aiohttp = "^3.11.13"
rapidfuzz = "^3.12.2"
spacy = "^3.8.4"
jsonlines = "^4.0.0"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
ruff = "^0.9.9"
mypy = "^1.15.0"
codespell = "^2.4.1"
isort = "^6.0.1"
types-requests = "^2.32.0.20250301"

[tool.mypy]
explicit_package_bases = true
ignore_missing_imports = true
files = ["src"]

[tool.codespell]
skip = ["./poetry.lock", "tmp*", "prompts"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
